CIV-904 refactor rename packages and views
