+++ src/main/routes/routes.ts
@@ -2,8 +2,8 @@ import homeRoute from './home';
 import unauthorisedRoute from './unauthorised';
 import dashboardRoute from './features/dashboard/index';
 import responseRoute from './features/response/index';
-import phoneDetailsRoute from './features/response/yourPhoneDetails/citizenPhoneController';
-import responseDobRoute from './features/response/yourDob/citizenDobController';
+import phoneDetailsRoute from './features/response/citizenPhoneNumber/citizenPhoneController';
+import responseDobRoute from './features/response/citizenDob/citizenDobController';
 
 export default [
   homeRoute,
